+++ src/super/i-block/i-block.ts
@@ -229,6 +229,18 @@ export default abstract class iBlock extends ComponentInterface {
 	readonly globalName?: string;
 
 	/**
+	 * Type of the component' root tag
+	 */
+	@prop(String)
+	readonly rootTag: string = 'div';
+
+	/**
+	 * Dictionary with additional attributes for the component' root tag
+	 */
+	@field()
+	rootAttrs: Dictionary = {};
+
+	/**
 	 * A component render cache key.
 	 * It's used to cache the component vnode.
 	 */