+++ src/super/i-block/CHANGELOG.md
@@ -9,6 +9,12 @@ Changelog
 > - :house:      [Internal]
 > - :nail_care:  [Polish]
 
+## v3.0.0-rc.??? (2021-??-??)
+
+#### :rocket: New Feature
+
+* Added new props `rootTag` and `rootAttrs`
+
 ## v3.0.0-rc.209 (2021-07-06)
 
 #### :bug: Bug Fix